+++ src/main/java/oshi/hardware/platform/linux/LinuxHardwareAbstractionLayer.java
@@ -19,10 +19,10 @@ package oshi.hardware.platform.linux;
 import oshi.hardware.CentralProcessor;
 import oshi.hardware.Display;
 import oshi.hardware.GlobalMemory;
-import oshi.hardware.HWDiskStore;
 import oshi.hardware.PowerSource;
 import oshi.hardware.Sensors;
 import oshi.hardware.common.AbstractHardwareAbstractionLayer;
+import oshi.hardware.common.HWDiskStore;
 import oshi.software.os.OSFileStore;
 import oshi.software.os.linux.LinuxFileSystem;