+++ src/main/java/oshi/hardware/platform/mac/MacDisks.java
@@ -13,25 +13,16 @@ package oshi.hardware.platform.mac;
 
 import java.util.ArrayList;
 import java.util.List;
-import javax.json.Json;
-import javax.json.JsonArrayBuilder;
-import javax.json.JsonBuilderFactory;
-import javax.json.JsonObject;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import oshi.hardware.Disks;
-import oshi.hardware.HWDiskStore;
-import oshi.json.NullAwareJsonObjectBuilder;
+
+import oshi.hardware.common.AbstractDisks;
+import oshi.hardware.common.HWDiskStore;
 
 /**
  * Mac hard disk implementation.
  *
  * @author enrico[dot]bianchi[at]gmail[dot]com
  */
-public class MacDisks implements Disks {
-
-    private JsonBuilderFactory jsonFactory = Json.createBuilderFactory(null);
-    private static final Logger LOG = LoggerFactory.getLogger(MacDisks.class);
+public class MacDisks extends AbstractDisks {
 
     @Override
     public HWDiskStore[] getDisks() {
@@ -39,18 +30,7 @@ public class MacDisks implements Disks {
 
         result = new ArrayList<>();
 
-        //TODO: extract disks hardware information
+        // TODO: extract disks hardware information
         return result.toArray(new HWDiskStore[result.size()]);
-    }
-
-    @Override
-    public JsonObject toJSON() {
-
-        JsonArrayBuilder array = jsonFactory.createArrayBuilder();
-
-        for (HWDiskStore store : getDisks()) {
-            array.add(store.toJSON());
-        }
-        return NullAwareJsonObjectBuilder.wrap(jsonFactory.createObjectBuilder()).add("disks", array).build();
     }
 }