+++ src/main/java/oshi/hardware/platform/windows/WindowsHardwareAbstractionLayer.java
@@ -19,11 +19,11 @@ package oshi.hardware.platform.windows;
 import oshi.hardware.CentralProcessor;
 import oshi.hardware.Display;
 import oshi.hardware.GlobalMemory;
-import oshi.hardware.HWDiskStore;
 import oshi.hardware.HardwareAbstractionLayer;
 import oshi.hardware.PowerSource;
 import oshi.hardware.Sensors;
 import oshi.hardware.common.AbstractHardwareAbstractionLayer;
+import oshi.hardware.common.HWDiskStore;
 import oshi.software.os.OSFileStore;
 import oshi.software.os.windows.WindowsFileSystem;